Create DraggingCD4020140428
This file creates a representation of CD40 from arrays and allows it to be dragged about. 
I can do it and I think I even understand it. 
There is probably a better way to do this than with the loops I have used but it least it seems to work.