Create cleaningCD40arraybyremoving20140429.html
This script will go to the Uniprot page for CD40 (hard coded). It will extract the XML file from there. It will extract the Features from the XML file. This Feature database will be checked against a list of notWantedFeatures. These notWantedFeatures will then be removed. 
The array that contains the data will then be written onto the webpage.